+++ src/control_plane/project_infrastructure/mod.rs
@@ -1,3 +1,4 @@
+mod plan_elasticsearch_project_resources;
 mod plan_meilisearch_project_resources;
 mod plan_opensearch_project_resources;
 mod plan_soketi_project_resources;
@@ -7,6 +8,7 @@ mod prepared_project_service;
 mod project_service_preparation_error;
 mod project_service_preparation_strategy;
 
+pub(crate) use plan_elasticsearch_project_resources::plan_elasticsearch_project_resources;
 pub(crate) use plan_meilisearch_project_resources::plan_meilisearch_project_resources;
 pub(crate) use plan_opensearch_project_resources::{
     opensearch_initial_admin_password, plan_opensearch_project_resources,